Finish title bar, link change, style change
Added lots to the title bar, made changes to the link appearance in the
link list to match the rest of the site for consistency, user pattern
recognition.  Disabled command mode.  Fixed issue at bottom of page with
border that acts as divider not going all the way down the page.  Still
need fo fix CSS a bit.